+++ gsi-server/src/main/java/de/gematik/idp/gsi/server/configuration/GsiConfiguration.java
@@ -1,5 +1,5 @@
 /*
- *  Copyright [2023] gematik GmbH
+ *  Copyright 2023 gematik GmbH
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -38,4 +38,5 @@ public class GsiConfiguration {
   private String fedmasterUrl;
   private KeyConfig sigKeyConfig;
   private KeyConfig tokenKeyConfig;
+  private String loglevel;
 }